Adjust types
