+++ packages/api/src/resolvers/queries/shopAnalytics.ts
@@ -1,0 +1,14 @@
+export default function shopInfo() {
+  const startOfToday = new Date();
+  startOfToday.setHours(0, 0, 0, 0);
+
+  const uptime = process.uptime();
+  const instanceStartTime = new Date(Date.now() - uptime * 1000);
+
+  return {
+    instanceStartTime,
+    startOfToday,
+  };
+}
+
+//
